backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
PyJWT[crypto]>=2.8.0
passlib[bcrypt]>=1.7.4
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.23
alembic>=1.12.1
PyPDF2>=3.0.1
python-docx>=1.1.0
google-genai>=0.3.0
python-dotenv>=1.0.0